ci: introduce workbench test namespace
