+++ workbench/app/LengthSpecification.php
@@ -2,20 +2,25 @@
 
 declare(strict_types=1);
 
-namespace Maartenpaauw\Specifications\Tests\Dummy;
+namespace Workbench\App;
 
 use Maartenpaauw\Specifications\CompositeSpecification;
 
 /**
  * @extends CompositeSpecification<string>
  */
-class UppercaseSpecification extends CompositeSpecification
+class LengthSpecification extends CompositeSpecification
 {
+    public function __construct(
+        private readonly int $length,
+    ) {
+    }
+
     /**
      * {@inheritDoc}
      */
     public function isSatisfiedBy(mixed $candidate): bool
     {
-        return strtoupper($candidate) === $candidate;
+        return strlen($candidate) === $this->length;
     }
 }
