+++ workbench/app/UppercaseSpecification.php
@@ -2,25 +2,20 @@
 
 declare(strict_types=1);
 
-namespace Maartenpaauw\Specifications\Tests\Dummy;
+namespace Workbench\App;
 
 use Maartenpaauw\Specifications\CompositeSpecification;
 
 /**
  * @extends CompositeSpecification<string>
  */
-class LengthSpecification extends CompositeSpecification
+class UppercaseSpecification extends CompositeSpecification
 {
-    public function __construct(
-        private readonly int $length,
-    ) {
-    }
-
     /**
      * {@inheritDoc}
      */
     public function isSatisfiedBy(mixed $candidate): bool
     {
-        return strlen($candidate) === $this->length;
+        return strtoupper($candidate) === $candidate;
     }
 }
